requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
Flask-SQLAlchemy==3.1.1
pydantic[email]==2.12.3
python-json-logger==4.0.0
ruff==0.14.3
flask-pydantic-spec==0.8.6
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
pymysql==1.1.2